+++ sync-core/src/main/java/com/cloudant/sync/replication/CouchDB.java
@@ -31,11 +31,6 @@ interface CouchDB {
     public String getIdentifier();
 
     /**
-     * Returns the name of the db
-     */
-    public String getDbName();
-
-    /**
      * Returns true if database exists, otherwise false
      */
     public boolean exists();