+++ sync-core/src/test/java/com/cloudant/sync/datastore/BasicDBObjectTest.java
@@ -15,7 +15,7 @@
 package com.cloudant.sync.datastore;
 
 import com.cloudant.sync.util.JSONUtils;
-import junit.framework.Assert;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;