+++ sync-core/src/test/java/com/cloudant/sync/datastore/DocumentRevisionBuilderTest.java
@@ -37,7 +37,7 @@ public class DocumentRevisionBuilderTest extends ReplicationTestBase {
         body.put("aKey","aValue");
         body.put("hello","world");
 
-        documentURI =  new URI(remoteDb.getCouchClient().getDefaultDBUri().toString()+"/someIdHere");
+        documentURI =  new URI(remoteDb.getCouchClient().getRootUri().toString()+"/someIdHere");
     }
 
 
@@ -209,7 +209,7 @@ public class DocumentRevisionBuilderTest extends ReplicationTestBase {
         attachments.put("bonsai-boston.jpg",bonsai);
         documentRev.put("_attachments",attachments);
 
-        URI uri =  new URI(remoteDb.getCouchClient().getDefaultDBUri().toString()+"/"+"someIdHere");
+        URI uri =  new URI(remoteDb.getCouchClient().getRootUri().toString()+"/"+"someIdHere");
 
         //create the document revision and test
 
@@ -264,7 +264,7 @@ public class DocumentRevisionBuilderTest extends ReplicationTestBase {
         attachments.put("bonsai-boston.jpg",bonsai);
         documentRev.put("_attachments",attachments);
 
-        URI uri =  new URI(remoteDb.getCouchClient().getDefaultDBUri().toString()+"/"+"someIdHere"+"?rev="+response.getRev());
+        URI uri =  new URI(remoteDb.getCouchClient().getRootUri().toString()+"/"+"someIdHere"+"?rev="+response.getRev());
 
         //create the document revision and test
 
@@ -303,7 +303,7 @@ public class DocumentRevisionBuilderTest extends ReplicationTestBase {
         remoteDoc.put("_attachments",remoteAttachments);
 
         remoteDb.create(remoteDoc);
-        documentURI = new URI(remoteDb.getCouchClient().getDefaultDBUri().toString()+"/"+"someIdHere");
+        documentURI = new URI(remoteDb.getCouchClient().getRootUri().toString()+"/"+"someIdHere");
 
         Map<String,Map<String,Object>> attachments = new HashMap<String,Map<String,Object>>();
         Map<String,Object> bonsai = createAttachmentMap(encodedAttachment,"text/plain",true);//new HashMap<String,Object>();