+++ sync-core/src/test/java/com/cloudant/sync/datastore/MultipartAttachmentWriterTests.java
@@ -16,7 +16,7 @@ package com.cloudant.sync.datastore;
 
 import com.cloudant.sync.util.TestUtils;
 
-import junit.framework.Assert;
+import org.junit.Assert;
 
 import org.junit.After;
 import org.junit.Before;