+++ sync-core/src/test/java/com/cloudant/sync/indexing/QueryBuilderTest.java
@@ -16,7 +16,8 @@ package com.cloudant.sync.indexing;
 
 
 import com.cloudant.sync.indexing.QueryBuilder;
-import junit.framework.Assert;
+
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;