+++ sync-core/src/test/java/com/cloudant/sync/replication/RemoteCheckpointDocTest.java
@@ -14,7 +14,7 @@
 
 package com.cloudant.sync.replication;
 
-import junit.framework.Assert;
+import org.junit.Assert;
 import org.junit.Test;
 
 public class RemoteCheckpointDocTest {